+++ src/sdf/array3d/imp.h
@@ -1,4 +1,4 @@
 struct Array3d;
-void array3d_ini(Array3d**, int x, int y, int z);
+void array3d_ini(Array3d**, size_t x, size_t y, size_t z);
 void array3d_fin(Array3d*);
-void array3d_copy(int x, int y, int z, float *D, /**/ Array3d*);
+void array3d_copy(size_t x, size_t y, size_t z, float *D, /**/ Array3d*);
